Return the original value of the add method
ActiveRecord has a feature called `verbose_query_logs` to display the source location of a SQL query. This is only triggered when the return value of the log method is truthy. The `add` method here changes the return value of the `add` method of `Logger` which it should not do.